Refactor world generation
