+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 2025-07-14
+- 1507 Break up worldGeneration.js into modular files
+
 ## 2025-06-28
 - 0100 Rig up running animation on Shift key press for animated player model.
 
@@ -14,9 +17,6 @@
 - 2309 Update remote players to run custom animations each frame
 - 2257 Add animated player replace button and improved options styling
 - 2242 Add options menu with asset downloader and glb analysis scripts
-- 2221 Document new agents directory and split instructions
-- 2218 Fade chat bubbles before hiding
-- 2214 Show sent chat immediately and keep log of messages
 
 ## Guidelines for future updates
 - List changes in reverse chronological order (newest first).
