+++ js/worldgen/constants.js
@@ -1,0 +1,16 @@
+export const ZONE_SIZE = 150;
+export const ZONES_PER_CHUNK_SIDE = 5;
+export const CHUNKS_PER_CLUSTER_SIDE = 5;
+
+export const CHUNK_SIZE = ZONE_SIZE * ZONES_PER_CHUNK_SIDE;
+export const CLUSTER_SIZE = CHUNK_SIZE * CHUNKS_PER_CLUSTER_SIDE;
+export const SPAWN_SAFE_RADIUS = 10;
+
+export const TERRAIN_AMPLITUDE = 10;
+export const TERRAIN_SCALE = 80;
+export const TERRAIN_TEXTURE_REPEAT_PER_ZONE = 50;
+export const TERRAIN_SEGMENTS = 250;
+export const BARRIERS_PER_ZONE = 0.3;
+export const PILLARS_PER_ZONE = 0.2;
+export const TREES_PER_ZONE = 0.5;
+export const CLOUD_COUNT = 100;
