+++ js/worldgen/random.js
@@ -1,0 +1,10 @@
+export class MathRandom {
+  constructor(seed) {
+    this.seed = seed;
+  }
+
+  random() {
+    const x = Math.sin(this.seed++) * 10000;
+    return x - Math.floor(x);
+  }
+}
